Improved linux recast installation

+++ Util/BuildTools/BuildUtilsDocker.sh
@@ -12,26 +12,25 @@ if [ -f "${FBX2OBJ_DIST}/FBX2OBJ" ]; then
   exit
 fi
 
-log "Downloading and preparing FBX SDK..."
-
 LIB_NAME=fbx202001_fbxsdk_linux
 FBXSDK_URL=https://www.autodesk.com/content/dam/autodesk/www/adn/fbx/2020-0-1/${LIB_NAME}.tar.gz
 
-echo "Downloading FBX SDK 2020"
+if [ ! -d "${FBX2OBJ_DEP_FOLDER}" ]; then
+  log "Downloading FBX SDK..."
+  wget -c "${FBXSDK_URL}" -P "${CARLA_DOCKER_UTILS_FOLDER}"
 
-wget -c "${FBXSDK_URL}" -P "${CARLA_DOCKER_UTILS_FOLDER}"
+  echo "Unpacking..."
+  mkdir -p "${FBX2OBJ_DEP_FOLDER}"
+  tar -xvzf "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}.tar.gz" -C "${CARLA_DOCKER_UTILS_FOLDER}" "${LIB_NAME}"
+  rm "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}.tar.gz"
 
-echo "Unpacking..."
-mkdir -p "${FBX2OBJ_DEP_FOLDER}"
-tar -xvzf "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}.tar.gz" -C "${CARLA_DOCKER_UTILS_FOLDER}" "${LIB_NAME}"
-rm "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}.tar.gz"
+  echo "Installing FBX SDK..."
+  echo -e "y\nyes\nn\n" | "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}" "${FBX2OBJ_DEP_FOLDER}"
+  echo
+  rm "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}"
+fi
 
-echo "Installing..."
-echo -e "y\nyes\nn\n" | "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}" "${FBX2OBJ_DEP_FOLDER}"
-echo
-rm "${CARLA_DOCKER_UTILS_FOLDER}/${LIB_NAME}"
-
-echo "Compiling FBX2OBJ..."
+log "Compiling FBX2OBJ..."
 mkdir -p "${FBX2OBJ_DIST}"
 mkdir -p "${FBX2OBJ_BUILD_FOLDER}"
 
@@ -41,9 +40,18 @@ cmake -G "Ninja" \
     -DCMAKE_CXX_FLAGS="-fPIC -std=c++14" \
     ..
 
+set +e
+
 ninja
 
+if [ $? -eq 1 ]; then
+  fatal_error "Make sure \"libxml2-dev\" is installed using:\n\n    sudo apt-get install libxml2-dev\n"
+  exit 1
+fi
+
 ninja install
+
+set -e
 
 popd >/dev/null
